backend/requirements.txt
# FastAPI and server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
authlib==1.3.0
httpx==0.26.0

# OpenAI
openai==1.12.0

# Validation and settings
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0

# Utilities
python-dotenv==1.0.1
aiofiles==23.2.1
Pillow==10.2.0
